requirements.txt
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
pandas>=2.0.0
openpyxl>=3.1.0
tqdm>=4.65.0 